requirements.txt
numpy
pandas
matplotlib
seaborn
lightkurve
tensorflow
scikit-learn
pymc3
batman-package
astropy
astroquery
arviz
